Minor changes to history and added more tests. Fixed a unhandled exception if command type happens to be null.

+++ src/logic/Logic.java
@@ -175,38 +175,42 @@ public class Logic {
 		Command.Type commandType = commandObject.getCommandType();
 		ArrayList<Task> userTasks = commandObject.getTasks();
 		ArrayList<String> argumentList = commandObject.getArguments();
-
-		switch (commandType) {
-			case ADD :
-				logger.info("ADD command detected");
-				return addItem(userTasks, argumentList, shouldPushToHistory,
-						isUndoHistory);
-			case DELETE :
-				logger.info("DELETE command detected");
-				return deleteItem(argumentList, shouldPushToHistory, isUndoHistory);
-			case EDIT :
-				logger.info("EDIT command detected");
-				return editItem(userTasks, argumentList, shouldPushToHistory,
-						isUndoHistory);
-			case DISPLAY :
-				logger.info("DISPLAY command detected");
-				return displayItems();
-			case UNDO :
-				logger.info("UNDO command detected");
-				return undoCommand();
-			case REDO :
-				logger.info("REDO command detected");
-				return redoCommand();
-			case SAVETO :
-				logger.info("SAVETO command detected");
-				return saveFilePath(argumentList);
-			case EXIT :
-				logger.info("EXIT command detected");
-				return exitProgram();
-			default :
+		if(commandType == null){
+			logger.warning("Command type is null!");
+			return ERROR_NO_COMMAND_HANDLER;
+		} else {
+			switch (commandType) {
+				case ADD :
+					logger.info("ADD command detected");
+					return addItem(userTasks, argumentList, shouldPushToHistory,
+							isUndoHistory);
+				case DELETE :
+					logger.info("DELETE command detected");
+					return deleteItem(argumentList, shouldPushToHistory, isUndoHistory);
+				case EDIT :
+					logger.info("EDIT command detected");
+					return editItem(userTasks, argumentList, shouldPushToHistory,
+							isUndoHistory);
+				case DISPLAY :
+					logger.info("DISPLAY command detected");
+					return displayItems();
+				case UNDO :
+					logger.info("UNDO command detected");
+					return undoCommand();
+				case REDO :
+					logger.info("REDO command detected");
+					return redoCommand();
+				case SAVETO :
+					logger.info("SAVETO command detected");
+					return saveFilePath(argumentList);
+				case EXIT :
+					logger.info("EXIT command detected");
+					return exitProgram();
+				default :
+					logger.warning("Command type cannot be identified!");
+					return ERROR_NO_COMMAND_HANDLER;
+			}
 		}
-		logger.warning("Command type cannot be identified!");
-		return ERROR_NO_COMMAND_HANDLER;
 	}
 
 	/**
